# openapi.yaml
openapi: 3.1.0
info:
  title: GMAO API
  description: API Proxy - Gateway entre frontend et données
  version: 2.2.1
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Proxy Directus /auth/login en conservant les cookies (session).
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Payload
              description: Identifiants Directus (email, password)
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError